fix: 修正小數點顯示錯誤 原本寫法導致小數點全部為0

+++ source/utils/number-to-decimal.js
@@ -3,9 +3,7 @@ import { toString, slice } from 'ramda';
 const numberToDecimal = numberStr => {
   if (isNaN(parseFloat(numberStr))) return '';
 
-  const decimal = Math.abs(parseFloat(numberStr) - +numberStr).toFixed(
-    numberStr.indexOf('.') > 0 ? parseFloat(numberStr.length - numberStr.indexOf('.') - 1) : 0
-  );
+  const decimal = Math.abs(parseInt(numberStr) - +numberStr).toFixed(numberStr.indexOf('.') > 0 ? parseInt(numberStr.length - numberStr.indexOf('.') - 1) : 0);
   return slice(2, Infinity)(decimal);
 };
